Netflix Clone This project is a simple yet visually appealing clone of the Netflix homepage, created using HTML and CSS. The clone features a hero section with a large background image and a call-to-action button, closely mimicking Netflix's engaging introductory layout. It includes a features section that highlights key benefits of the service, using images and descriptive text. The FAQ section provides a list of frequently asked questions with a clean, interactive design, allowing users to easily find answers. The footer is styled to include additional links and contact information, echoing the structure and design of Netflix's original footer. The project utilizes responsive design principles to ensure a seamless experience across various devices and screen sizes, reflecting a modern, dark-themed aesthetic similar to Netflix's interface.
Netflix Clone
This project is a simple yet visually appealing clone of the Netflix homepage, created using HTML and CSS. The clone features a hero section with a large background image and a call-to-action button, closely mimicking Netflix's engaging introductory layout. It includes a features section that highlights key benefits of the service, using images and descriptive text. The FAQ section provides a list of frequently asked questions with a clean, interactive design, allowing users to easily find answers. The footer is styled to include additional links and contact information, echoing the structure and design of Netflix's original footer. The project utilizes responsive design principles to ensure a seamless experience across various devices and screen sizes, reflecting a modern, dark-themed aesthetic similar to Netflix's interface.